requirements.txt
absl-py==2.3.0
ale-py==0.11.1
box2d-py==2.3.8
cffi==1.17.1
chex==0.1.89
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
etils==1.12.2
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.18.0
flax==0.10.6
fonttools==4.58.4
fsspec==2025.5.1
glfw==2.9.0
grpcio==1.73.1
grpcio-tools==1.73.1
gymnasium==1.1.1
humanize==4.12.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
iniconfig==2.1.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
mujoco==3.3.3
mujoco-py==2.1.2.14
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.16
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
proglog==0.1.12
protobuf==6.31.1
pycparser==2.22
pygame==2.6.1
Pygments==2.19.2
PyOpenGL==3.1.9
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
rich==14.0.0
scipy==1.16.0
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
swig==4.3.1
sympy==1.14.0
tensorstore==0.1.75
toolz==1.0.0
torch==2.7.1
tqdm==4.67.1
treescope==0.1.9
typing_extensions==4.14.0
zipp==3.23.0
